Reverted namespacing
