+++ lib/currency_cloud/account.rb
@@ -1,0 +1,8 @@
+module CurrencyCloud
+  class Account
+    include CurrencyCloud::Resource
+
+    resource :accounts
+    actions :create, :retrieve, :find, :update, :current
+  end
+end
