+++ lib/currency_cloud/actions.rb
@@ -1,0 +1,13 @@
+require_relative "actions/create"
+require_relative "actions/current"
+require_relative "actions/delete"
+require_relative "actions/find"
+require_relative "actions/instance_delete"
+require_relative "actions/retrieve"
+require_relative "actions/save"
+require_relative "actions/update"
+
+module CurrencyCloud
+  module Actions
+  end
+end
