+++ lib/currency_cloud/contact.rb
@@ -1,0 +1,8 @@
+module CurrencyCloud
+  class Contact
+    include CurrencyCloud::Resource
+
+    resource :contacts
+    actions :create, :retrieve, :find, :update, :current
+  end
+end
