+++ lib/currency_cloud/conversion.rb
@@ -1,0 +1,8 @@
+module CurrencyCloud
+  class Conversion
+    include CurrencyCloud::Resource
+
+    resource :conversions
+    actions :create, :retrieve, :find
+  end
+end
