+++ lib/currency_cloud/conversion_dates.rb
@@ -1,0 +1,5 @@
+module CurrencyCloud
+  class ConversionDates
+    include CurrencyCloud::Resource
+  end
+end
