+++ lib/currency_cloud/currency.rb
@@ -1,0 +1,5 @@
+module CurrencyCloud
+  class Currency
+    include CurrencyCloud::Resource
+  end
+end
