+++ lib/currency_cloud/pagination.rb
@@ -1,7 +1,4 @@
 module CurrencyCloud
-  
   class Pagination < OpenStruct
-    
   end
-  
 end
