+++ lib/currency_cloud/payer.rb
@@ -1,0 +1,8 @@
+module CurrencyCloud
+  class Payer
+    include CurrencyCloud::Resource
+
+    resource :payers
+    actions :retrieve
+  end
+end
