+++ lib/currency_cloud/payment.rb
@@ -1,0 +1,8 @@
+module CurrencyCloud
+  class Payment
+    include CurrencyCloud::Resource
+
+    resource :payments
+    actions :create, :retrieve, :find, :delete, :update
+  end
+end
