+++ lib/currency_cloud/rates.rb
@@ -1,0 +1,5 @@
+module CurrencyCloud
+  class Rates
+    include CurrencyCloud::Resource
+  end
+end
