+++ lib/currency_cloud/request_handler.rb
@@ -1,20 +1,18 @@
 module CurrencyCloud
-
   class RequestHandler
-    
     attr_reader :session
-    
+
     def initialize(session = CurrencyCloud.session)
       @session = session
     end
-    
+
     def get(route, params={}, opts={})
       retry_authenticate('get', route, params, opts) do |url, params, options|
         options.merge!(:query => params)
         HTTParty.get(url, options)
       end
     end
-    
+
     def post(route, params={}, opts={})
       retry_authenticate('post', route, params, opts) do |url, params, options|
         options.merge!(:body => params)
@@ -23,9 +21,10 @@ module CurrencyCloud
     end
 
     private
+
     def retry_authenticate(verb, route, params, opts)
       should_retry = opts[:should_retry].nil? ? true : opts.delete(:should_retry)
-      
+
       params = process_params(params)
       full_url = build_url(route)
 
@@ -56,9 +55,9 @@ module CurrencyCloud
 
     def process_params(params)
       if session && session.on_behalf_of && CurrencyCloud::UUID_REGEX.match(session.on_behalf_of)
-        params.merge!(on_behalf_of: session.on_behalf_of)  
+        params.merge!(on_behalf_of: session.on_behalf_of)
       end
-      
+
       params
     end
 
@@ -67,9 +66,9 @@ module CurrencyCloud
       headers['X-Auth-Token'] = session.token if session && session.token
       headers
     end
-        
+
     def build_url(route)
-      "#{session.environment_url}/#{CurrencyCloud::ApiVersion}/" + route
+      "#{session.environment_url}/#{CurrencyCloud::API_VERSION}/" + route
     end
   end
 end
