--- lib/currency_cloud/resources/account.rb
@@ -1,10 +1,0 @@
-module CurrencyCloud
-  module Resources
-    class Account
-      include CurrencyCloud::Resource
-
-      resource :accounts
-      actions :create, :retrieve, :find, :update, :current
-    end
-  end
-end
