--- lib/currency_cloud/resources/balance.rb
@@ -1,14 +1,0 @@
-module CurrencyCloud
-  module Resources
-    class Balance
-      include CurrencyCloud::Resource
-
-      resource :balances
-      actions :find
-
-      def self.currency(ccy)
-        new(client.get(ccy))
-      end
-    end
-  end
-end
