--- lib/currency_cloud/resources/beneficiary.rb
@@ -1,14 +1,0 @@
-module CurrencyCloud
-  module Resources
-    class Beneficiary
-      include CurrencyCloud::Resource
-
-      resource :beneficiaries
-      actions :create, :retrieve, :find, :update, :delete
-
-      def self.validate(params)
-        new(client.post("validate", params))
-      end
-    end
-  end
-end
