--- lib/currency_cloud/resources/contact.rb
@@ -1,10 +1,0 @@
-module CurrencyCloud
-  module Resources
-    class Contact
-      include CurrencyCloud::Resource
-
-      resource :contacts
-      actions :create, :retrieve, :find, :update, :current
-    end
-  end
-end
