--- lib/currency_cloud/resources/conversion.rb
@@ -1,10 +1,0 @@
-module CurrencyCloud
-  module Resources
-    class Conversion
-      include CurrencyCloud::Resource
-
-      resource :conversions
-      actions :create, :retrieve, :find
-    end
-  end
-end
