--- lib/currency_cloud/resources/conversion_dates.rb
@@ -1,7 +1,0 @@
-module CurrencyCloud
-  module Resources
-    class ConversionDates
-      include CurrencyCloud::Resource
-    end
-  end
-end
