--- lib/currency_cloud/resources/currency.rb
@@ -1,7 +1,0 @@
-module CurrencyCloud
-  module Resources
-    class Currency
-      include CurrencyCloud::Resource
-    end
-  end
-end
