--- lib/currency_cloud/resources/payer.rb
@@ -1,10 +1,0 @@
-module CurrencyCloud
-  module Resources
-    class Payer
-      include CurrencyCloud::Resource
-
-      resource :payers
-      actions :retrieve
-    end
-  end
-end
