--- lib/currency_cloud/resources/payment.rb
@@ -1,10 +1,0 @@
-module CurrencyCloud
-  module Resources
-    class Payment
-      include CurrencyCloud::Resource
-
-      resource :payments
-      actions :create, :retrieve, :find, :delete, :update
-    end
-  end
-end
