--- lib/currency_cloud/resources/rates.rb
@@ -1,7 +1,0 @@
-module CurrencyCloud
-  module Resources
-    class Rates
-      include CurrencyCloud::Resource
-    end
-  end
-end
