--- lib/currency_cloud/resources/settlement_account.rb
@@ -1,7 +1,0 @@
-module CurrencyCloud
-  module Resources
-    class SettlementAccount
-      include CurrencyCloud::Resource
-    end
-  end
-end
