--- lib/currency_cloud/resources/transaction.rb
@@ -1,10 +1,0 @@
-module CurrencyCloud
-  module Resources
-    class Transaction
-      include CurrencyCloud::Resource
-
-      resource :transactions
-      actions :retrieve, :find
-    end
-  end
-end
