+++ lib/currency_cloud/response_handler.rb
@@ -1,16 +1,14 @@
 module CurrencyCloud
-  
   class ResponseHandler
-    
     attr_reader :verb, :route, :params, :response
-    
+
     def initialize(verb, route, params, response)
       @verb = verb
       @route = route
       @params = params
       @response = response
     end
-    
+
     def process
       if success?
         return parsed_response
@@ -18,13 +16,13 @@ module CurrencyCloud
         handle_failure
       end
     end
-    
+
     private
-    
+
     def success?
       [200, 202].include?(response.code)
     end
-    
+
     def handle_failure
       error_class = case response.code
         when 400 then BadRequestError
@@ -37,7 +35,7 @@ module CurrencyCloud
       raise error_class.new(verb, route, params, response) if error_class
       raise UnexpectedError.new(verb, route, params, response)
     end
-    
+
     def parsed_response
       @parsed_response ||= JSON.parse(response.body)
     end
