+++ lib/currency_cloud/settlement_account.rb
@@ -1,0 +1,5 @@
+module CurrencyCloud
+  class SettlementAccount
+    include CurrencyCloud::Resource
+  end
+end
