+++ lib/currency_cloud/transaction.rb
@@ -1,0 +1,8 @@
+module CurrencyCloud
+  class Transaction
+    include CurrencyCloud::Resource
+
+    resource :transactions
+    actions :retrieve, :find
+  end
+end
