+++ spec/integration/authentication_spec.rb
@@ -19,7 +19,7 @@ describe 'Authentication', :vcr => true do
     CurrencyCloud.api_key = nil
     CurrencyCloud.token = '7fbba909f66ee6721b2e20a5fa1ccae7'
 
-    response = CurrencyCloud::Resources::Beneficiary.find
+    response = CurrencyCloud::Beneficiary.find
     expect(response).to_not be_nil
   end
 
@@ -31,7 +31,7 @@ describe 'Authentication', :vcr => true do
   it 'handles session timeout error' do
     CurrencyCloud.token = '3907f05da86533710efc589d58f51f45'
 
-    response = CurrencyCloud::Resources::Beneficiary.find
+    response = CurrencyCloud::Beneficiary.find
     expect(response).to_not be_nil
 
     #should have changed after reauthentication
